requirements.txt
asgiref==3.9.1
boto3==1.34.17
botocore==1.34.162
dj-database-url==2.1.0
Django==5.2.6
django-cors-headers==4.7.0
django-storages==1.14.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
s3transfer==0.10.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.6.0
